rename ImageViewer to SiteDetails
